+++ fastai/imports/core.py
@@ -12,7 +12,7 @@ from copy import copy, deepcopy
 from dataclasses import dataclass, field
 from enum import Enum, IntEnum
 from fastprogress import master_bar, progress_bar
-from functools import partial, reduce
+from functools import partial, reduce, singledispatch
 from IPython.core.debugger import set_trace
 from matplotlib import patches, patheffects
 from numpy import array, cos, exp, log, sin, tan, tanh